+++ .gitignore
@@ -23,3 +23,4 @@
 *.x86_64
 *.hex
 fansi
+examples/*
