+++ include/util.h
@@ -1,7 +1,10 @@
 #ifndef UTIL_H
 #define UTIL_H
 
+#include <ctype.h>
+
 // prototypes
 void print_usage();
+char *trimwhitespace(char *str);
 
 #endif // UTIL_H
